chore(migrations): change relationship integers to bigints in migrations
This won't apply to existing applications who will need to handle this
migration for themselves if needed. However, it will make it easier for
new applications.

Closes #30

+++ db/migrate/20141012174250_create_authie_sessions.rb
@@ -4,7 +4,7 @@ class CreateAuthieSessions < ActiveRecord::Migration[6.1]
   def change
     create_table :authie_sessions do |t|
       t.string :token, :browser_id
-      t.integer :user_id
+      t.bigint :user_id
       t.boolean :active, default: true
       t.text :data
       t.datetime :expires_at
