+++ db/migrate/20150109144120_add_parent_id_to_authie_sessions.rb
@@ -2,6 +2,6 @@
 
 class AddParentIdToAuthieSessions < ActiveRecord::Migration[6.1]
   def change
-    add_column :authie_sessions, :parent_id, :integer
+    add_column :authie_sessions, :parent_id, :bigint
   end
 end
